requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
python-nmap
jinja2
pydantic
aiofiles
